Group all config file related functionality into "schemas".
