+++ square/schemas.py
@@ -9,7 +9,12 @@ This module defines utility functions to define these filters.
 """
 import copy
 import logging
-from typing import List, Union
+from types import SimpleNamespace
+from typing import List, Tuple, Union
+
+import pydantic
+import yaml
+from square.dtypes import Config, Filepath
 
 # Convenience.
 logit = logging.getLogger("square")
@@ -101,3 +106,46 @@ def merge(src: list, dst: list) -> list:
 
     _update(src, dst)
     return dst
+
+
+def load_config(fname: Filepath) -> Tuple[Config, bool]:
+    """Parse the Square configuration file `fname` and return it as a `Config`."""
+    err_resp = Config(Filepath(""), kubeconfig=Filepath(""), kubecontext=None), True
+    fname = Filepath(fname)
+
+    # Load the configuration file.
+    try:
+        raw = yaml.safe_load(Filepath(fname).read_text())
+    except FileNotFoundError as e:
+        logit.error(f"Cannot load config file <{fname}>: {e.args[1]}")
+        return err_resp
+    except yaml.YAMLError as exc:
+        msg = f"Could not parse YAML file {fname}"
+
+        # Special case: parser supplied location information.
+        mark = getattr(exc, "problem_mark", SimpleNamespace(line=-1, column=-1))
+        line, col = (mark.line + 1, mark.column + 1)
+        msg = f"YAML format error in {fname}: Line {line} Column {col}"
+        logit.error(msg)
+        return err_resp
+
+    if not isinstance(raw, dict):
+        logit.error(f"Config file <{fname}> has invalid structure")
+        return err_resp
+
+    # Parse the configuration into `ConfigFile` structure.
+    try:
+        cfg = Config(**raw)
+    except (pydantic.ValidationError, TypeError) as e:
+        logit.error(f"Schema is invalid: {e}")
+        return err_resp
+
+    # Remove the "_common_" filter and merge it into all the other filters.
+    common = cfg.filters.pop("_common_", [])
+    cfg.filters = {k: merge(common, v) for k, v in cfg.filters.items()}
+
+    # Make the necessary adjustments where the values in the file do not
+    # translate verbatim.
+    cfg.folder = fname.parent.absolute() / cfg.folder
+    cfg.selectors.kinds = set(cfg.selectors.kinds)
+    return cfg, False
