+++ tests/test_main.py
@@ -9,6 +9,7 @@ import pytest
 import square.k8s as k8s
 import square.main as main
 import square.manio as manio
+import square.schemas as schemas
 import square.square as square
 import yaml
 from square.dtypes import (
@@ -49,7 +50,7 @@ def fname_param_config(tmp_path) -> Generator[
     del ref
 
     # Load the sample configuration.
-    config, err = square.load_config(fname_square)
+    config, err = schemas.load_config(fname_square)
     assert not err
 
     # Point the folder and kubeconfig to temporary versions.
